Remove unnecessary cast

+++ xswap/src/xswap_wrapper.cpp
@@ -87,7 +87,7 @@ static PyObject* wrap_xswap(PyObject *self, PyObject *args) {
     stats.num_swaps = num_swaps;
 
     // Perform XSwap
-    swap_edges(edges, num_swaps, valid_cond, &stats, (unsigned long long int)max_malloc);
+    swap_edges(edges, num_swaps, valid_cond, &stats, max_malloc);
 
     // Get new edges as python list
     PyObject* py_list = edges_to_py_list(edges);
